// src/modules/auth/dto/reset-password.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class ResetPasswordDto {
  @ApiProperty({
    description: 'New password of the user',
    example: 'Password@123',
    type: String,
  })
  password: string;

  @ApiProperty({
    description: 'Hash generated by backend',
    type: String,
  })
  hash: string;
}